+++ CHANGELOG.rst
@@ -1,9 +1,23 @@
 Change Log
 ===========
 
-2.0.dev2
------
+2.1.0
+------
+* Added programs and docs for `fracsurvive`.
+
+* Added ``--scalebar`` to ``dms2_logoplot``.
+
+* Add `grouplabel` option and preserve group order for faceted plots by batch programs.
+
+* Handle dependencies without `__version__` attribute
+
+2.0.2
+------
+* Added ``--sitemask`` option to ``dms2_bcsubamp`` / ``dms2_batch_bcsubamp``.
+
 * Standardized color scheme in ``*_cumulmutcounts.pdf`` plot.
+
+* Ensure naturally sorted average prefs from ``dms2_batch_prefs``.
 
 2.0.1
 ------